requirements.txt
Flask==2.0.1
Pillow==7.0.0
tesseract==0.1.3
gunicorn==20.1.0
pytesseract==0.3.8
